Fix login exploit
When the attacker knows the victim has just joined a server, they would get a 2-minute window to attempt joining the same server with the victim's username, and would succeed (the victim would see "Logged in from another location" screen).

The fix makes severId not based on server's IP, but instead on player's IP, which makes the exploit impossible to execute without having access to the victim's network.